Fix creating multiple YouTube players.
The youtubePlayer factory function tries to load the YouTube API
every time it's called. However the YouTube API only calls its
load event (the global callback) once, when it's first loaded.
As a result, the youtubeIframeAPI promise never resolves after the
first call.

This patch only loads the iframe API on the first call, and reuses
the promise on subsequent calls. (Similar to what it did before
1a01ff9cc)